added refactoring and java doc changes
