+++ src/main/java/edu/northeastern/truthtree/common/normalization/INormalizationStrategy.java
@@ -1,9 +1,9 @@
-package edu.northeastern.truthtree.common.normalizationStrategy;
+package edu.northeastern.truthtree.common.normalization;
 
 import java.util.List;
 
 /**
- * Represents
+ * Represents interface for NormalizationStrategy followed in the application.
  */
 public interface INormalizationStrategy {
   public List normalize(List result, List normalizationParameters);
