+++ src/main/java/edu/northeastern/truthtree/common/normalization/NormalizationStrategyFactory.java
@@ -1,4 +1,4 @@
-package edu.northeastern.truthtree.common.normalizationStrategy;
+package edu.northeastern.truthtree.common.normalization;
 
 import edu.northeastern.truthtree.enums.NormalizationType;
