+++ src/main/java/edu/northeastern/truthtree/common/normalization/PerCapitaNormalizationStrategy.java
@@ -1,0 +1,32 @@
+package edu.northeastern.truthtree.common.normalization;
+
+import java.util.List;
+
+import edu.northeastern.truthtree.AppConst;
+
+/**
+ * This strategy divides attribute values with the population of the location for the given year.
+ */
+public class PerCapitaNormalizationStrategy extends AbstractNormalizationStrategy {
+
+  /**
+   * Initializes parameters map with the normalization parameters meant for per-capita normalization
+   * strategy.
+   *
+   * @param normalizationParameters represents a list containing values required for normalization.
+   */
+  @Override
+  protected void setParameterMap(List normalizationParameters) {
+    super.updateParameterMap(normalizationParameters, AppConst.POPULATION_PARAMETER);
+  }
+
+  /**
+   * Applies normalization with on the result and given type.
+   * @param result represents a list containing values to be normalized.
+   * @return
+   */
+  @Override
+  protected List apply(List result) {
+    return super.updateNormalizationValues(result, AppConst.PER_CAPITA);
+  }
+}
